+++ frontend/app/src/modules/dashboard/graph/types.ts
@@ -1,0 +1,5 @@
+import type { NetValue } from '@rotki/common';
+
+export interface NetValueChartData extends NetValue {
+  snapshotCount: number;
+}